added image controller
