--- app/imageUpload/image.jade
@@ -1,7 +1,0 @@
-extends ../templates/layout
-
-block content
-  h1 Image Upload
-  form(action="/image", method="post", enctype="multipart/form-data")
-    input(type='file' name='image' accept="image/*")
-    input(type="submit" value="Upload")
