+++ app/post/index.jade
@@ -1,8 +1,11 @@
 extends ../templates/layout
 
 block content
-  form(action='/post', method='POST')
+  form(action='/post', method='POST', enctype="multipart/form-data")
+    label(for='text') Message:
     input(name='text')
+    label(for='image') Upload Image:
+    input(type='file' name='image' accept="image/*")
     input(type='submit')
 
   ul
