+++ app/routes.js
@@ -4,11 +4,9 @@ var express = require('express');
 var router = express.Router();
 
 var home = require('./home/routes');
-var image = require('./imageUpload/routes');
 var post = require('./post/routes');
 
 router.use('/', home);
-router.use('/image', image);
 router.use('/', post);
 
 module.exports = router;
